feat: restart mpd on daily restart
Sometimes mpd seems to get stuck. Restart it in order to be safe.

Restarting it resumes the old state of mpd, playing or not playing.
So restart while AirPlay is running will start playing in parallel.
This seems like a not that often situation at that time so this is
currently ok.